Added LockProvider

+++ src/main/java/project/models/ModelImpl.java
@@ -1,11 +1,15 @@
 package project.models;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import project.data.DataStorage;
 import project.exceptions.TaskNotFoundException;
 import project.exceptions.UserExistsException;
 import project.exceptions.UserNotFoundException;
+import project.locks.AutoCloseableReentrantLock;
+import project.locks.LocksProvider;
 
 import java.util.List;
 import java.util.Objects;
@@ -13,11 +17,16 @@ import java.util.Objects;
 
 @Component
 public class ModelImpl implements Model {
+    private static final Logger log = LoggerFactory.getLogger(ModelImpl.class);
     private DataStorage dataStorage;
+    private LocksProvider locksProvider;
+    private static volatile boolean sleep = true;
 
     @Autowired
-    public ModelImpl(DataStorage dataStorage) {
+    public ModelImpl(DataStorage dataStorage,
+                     LocksProvider locksProvider) {
         this.dataStorage = dataStorage;
+        this.locksProvider = locksProvider;
     }
 
     @Override
@@ -48,20 +57,41 @@ public class ModelImpl implements Model {
     }
 
     @Override
-    public Task updateTask(long taskId, TaskData taskData) throws TaskNotFoundException {
+    public Task updateTask(long taskId, TaskData taskData) throws Exception {
         if (!dataStorage.getTaskById(taskId).isPresent()) throw new TaskNotFoundException();
-        return dataStorage.updateTask(taskId, taskData);
+        try(final AutoCloseableReentrantLock lock = locksProvider.provideTaskLock(taskId).open()) {
+            if (sleep) {
+                log.info("Спим 20 секунд");
+                sleep = false;
+                try {
+                    Thread.sleep(20000);
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            } else {
+                log.info("Не спим");
+                sleep = true;
+            }
+            return dataStorage.updateTask(taskId, taskData);
+        }
+    }
+
+
+    @Override
+    public void deleteTask(long taskId) throws Exception {
+        if (!dataStorage.getTaskById(taskId).isPresent()) throw new TaskNotFoundException();
+        try(final AutoCloseableReentrantLock lock = locksProvider.provideTaskLock(taskId).open()) {
+            log.info("Task to delete: " + dataStorage.getTaskById(taskId));
+            dataStorage.deleteTask(taskId);
+        }
     }
 
     @Override
-    public void deleteTask(long taskId) throws TaskNotFoundException {
-        if (!dataStorage.getTaskById(taskId).isPresent()) throw new TaskNotFoundException();
-        dataStorage.deleteTask(taskId);
-    }
-
-    @Override
-    public void deleteUser(long userId) throws UserNotFoundException {
-        dataStorage.deleteUser(userId);
+    public void deleteUser(long userId) throws Exception {
+        if (!dataStorage.getUser(userId).isPresent()) throw new UserNotFoundException();
+        try (final AutoCloseableReentrantLock lock = locksProvider.provideUserLock(userId)){
+            dataStorage.deleteUser(userId);
+        }
     }
 
     @Override
@@ -75,6 +105,5 @@ public class ModelImpl implements Model {
         if (!dataStorage.getUser(userId).isPresent()) throw new UserNotFoundException();
         return dataStorage.getAllActiveTask(userId);
     }
-
 
 }
